refactor(poa): move `process_reciepts()` to crate::utils
