+++ crates/consensus/authority/src/block_fetcher.rs
@@ -94,7 +94,14 @@ where
                 let sealed_block_with_senders =
                     SealedBlockWithSenders::new(sealed_block, senders).expect("senders are valid");
                 // Process Botanix specific logs
-                match self.process_reciepts(&bundle_state, false).await {
+                match crate::utils::process_reciepts(
+                    &self.bitcoin_block_source,
+                    &mut self.btc_server.clone(),
+                    &bundle_state,
+                    false,
+                )
+                .await
+                {
                     Ok(_) => {}
                     Err(e) => {
                         error!(target: "consensus::authority", ?e, "Failed to process botanix log");
